+++ sources/@roots/bud-eslint/tsconfig.json
@@ -8,8 +8,7 @@
   "exclude": ["lib", "node_modules", "**/*.test.ts"],
   "references": [
     {"path": "./../bud-framework/tsconfig.json"},
-    {"path": "./../bud-api/tsconfig.json"},
-    {"path": "./../bud-build/tsconfig.json"},
+    {"path": "./../bud/tsconfig.json"},
     {"path": "./../bud-support/tsconfig.json"}
   ]
 }
