+++ sources/@roots/bud-prettier/package.json
@@ -63,7 +63,7 @@
     "@types/node": "16.18.2"
   },
   "dependencies": {
-    "@roots/bud": "workspace:*",
+    "@roots/bud": "workspace:sources/@roots/bud",
     "@roots/bud-framework": "workspace:sources/@roots/bud-framework",
     "@roots/bud-support": "workspace:sources/@roots/bud-support",
     "prettier": "2.7.1",
