+++ sources/@roots/bud-tailwindcss/src/types.ts
@@ -1,0 +1,11 @@
+import type BudTailwindCss from './extension.js'
+
+declare module '@roots/bud-framework' {
+  interface Bud {
+    tailwind: BudTailwindCss
+  }
+
+  interface Modules {
+    '@roots/bud-tailwindcss': BudTailwindCss
+  }
+}
