+++ sources/@roots/bud-typescript/package.json
@@ -65,7 +65,6 @@
   },
   "devDependencies": {
     "@jest/globals": "29.2.2",
-    "@roots/bud-api": "workspace:sources/@roots/bud-api",
     "@skypack/package-check": "0.2.2",
     "@types/fork-ts-checker-webpack-plugin": "0.4.5",
     "@types/node": "16.18.2",
